Open Router/openrouter/requirements.txt
openai
openai-agents
python-dotenv
nest_asyncio
